refactor: sidebar directory
