+++ src/components/editor/sidebar/sidebar-element-settings-tab/logo-banner-setting.tsx
